Suppliers panel rekork finish

+++ frontend/maripos/src/components/SettingsPanel/SettingsCategories/Products/Options/Supliers/SettingsOption_Supliers.tsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react";
 import {SupliersList, SubmitPasswordResponse} from "../../../../../../interfaces/interfaces";
 import SuplierListItem from "./SuplierListItem";
 import SubmitPassword from "../../../../../SubmitPassword/SubmitPassword";
+import EditSuplier from "./EditSuplier";
 
 function SettingsOption_Supliers(){
     const [list, setList] = useState<SupliersList[]>([]);
@@ -52,11 +53,20 @@ function SettingsOption_Supliers(){
             setItemSelected(selected);
     }
 
-    function insertUser(){
+    function insertSuplier(){
         let newUser:SupliersList = {"Descripcion":"", "RFC":"", "_id":{"$oid":"new"}};
         setList((prevState)=>[...prevState, newUser]);
         setAction(`${process.env.REACT_APP_ACTIONS_NEW_SUPLIER}`);
         setItemSelected("");
+    }
+
+    function editSuplier(){
+        setAction(`${process.env.REACT_APP_ACTIONS_MODIFY_SUPLIER}`);
+    }
+
+    function deleteSuplier(){
+        setAction(`${process.env.REACT_APP_ACTIONS_DELETE_SUPLIER}`);
+        submit();
     }
 
     function cancel(){
@@ -75,38 +85,33 @@ function SettingsOption_Supliers(){
         let passwordButton = document.getElementById(passwordId["dismiss"]) as HTMLButtonElement;
         let newNameInput = document.getElementById("NewSuplierInput") as HTMLInputElement;
         let newRFCinput = document.getElementById("NewRFCInput") as HTMLInputElement;
-        let newSuplierObject = [{"Descripcion":newNameInput.value,
-                                "RFC":newRFCinput.value}];
+        let newSuplierObject = [{"Descripcion":newNameInput?.value,
+                                "RFC":newRFCinput?.value}];
 
                                 
-        await fetch(`${process.env.REACT_APP_BACKEND}/provider`,{
-            method:"POST",
-            headers:{'Content-Type':'application/json'},
-            body:JSON.stringify(newSuplierObject)
-        });
+        if(action===`${process.env.REACT_APP_ACTIONS_NEW_SUPLIER}`){
+            await fetch(`${process.env.REACT_APP_BACKEND}/provider`,{
+                method:"POST",
+                headers:{'Content-Type':'application/json'},
+                body:JSON.stringify(newSuplierObject)
+            });
+        }
+        if(action===`${process.env.REACT_APP_ACTIONS_MODIFY_SUPLIER}`){
+            await fetch(`${process.env.REACT_APP_BACKEND}/provider/${itemSelected}`,{
+                method:"PUT",
+                headers:{'Content-Type':'application/json'},
+                body:JSON.stringify(newSuplierObject)
+            });
+        }
+        if(action===`${process.env.REACT_APP_ACTIONS_DELETE_SUPLIER}`){
+            await fetch(`${process.env.REACT_APP_BACKEND}/provider/${itemSelected}`, {method:"DELETE"});
+        }
 
         resetSupliersList();
         passwordButton.click();
     }
     
-    
-    
-    
-    
-    
-    function deleteUser(){
-        let passwordButton = document.getElementById(passwordId["toggler"]) as HTMLButtonElement;
-        passwordButton.click();
-    }
-
-    async function deleteSuccessful(response:SubmitPasswordResponse){
-        let passwordButton = document.getElementById(passwordId["dismiss"]) as HTMLButtonElement;
-        
-        await fetch(`${process.env.REACT_APP_BACKEND}/provider/${itemSelected}`, {method:"DELETE"});
-        passwordButton.click();
-    }
-
-    function passwordFailure(response:SubmitPasswordResponse){
+    function submitFailure(response:SubmitPasswordResponse){
         alert(`ERROR: ${response["message"]}`);
     }
 
@@ -115,19 +120,35 @@ function SettingsOption_Supliers(){
         passwordClose.click();
     }
 
-    function checkValidSuplier(input:string){
+    function checkValidSuplier(input:string, rfc:string){
         
         for(let i=0; i<list.length;i++){
             let previous = list[i]["Descripcion"].toLowerCase().replaceAll(" ","");
+            let previousRFC = list[i]["RFC"].toLowerCase().replaceAll(" ","");
+
             if(input===""){
                 setValidSuplier(false);
                 setValidMessage("Introduce un nombre");
                 return;
             }
             if(input===previous){
-                setValidSuplier(false);
-                setValidMessage("Este proveedor ya existe");
-                return;
+                if(action===`${process.env.REACT_APP_ACTIONS_NEW_SUPLIER}`){
+                    setValidSuplier(false);
+                    setValidMessage("Este proveedor ya existe");
+                    return;
+                }
+                if(action===`${process.env.REACT_APP_ACTIONS_MODIFY_SUPLIER}`){
+                    if(list[i]["_id"]["$oid"] !== itemSelected){
+                        setValidSuplier(false);
+                        setValidMessage("Este proveedor ya existe");
+                        return;
+                    }
+                    else if(rfc === previousRFC){
+                        setValidSuplier(false);
+                        setValidMessage("Este proveedor ya existe");
+                        return;
+                    }
+                }
             }
         }
         setValidSuplier(true);
@@ -159,7 +180,7 @@ function SettingsOption_Supliers(){
                                         <tbody>
                                             {list.map((item)=>{
                                                 return (
-                                                    <SuplierListItem key={item["_id"]["$oid"]} item={item} onSelect={selectItem} selected={itemSelected} userInputCheck={checkValidSuplier}/>
+                                                    <SuplierListItem key={item["_id"]["$oid"]} item={item} onSelect={selectItem} selected={itemSelected} userInputCheck={checkValidSuplier} action={action}/>
                                                 );
                                             })}
                                         </tbody>
@@ -173,9 +194,9 @@ function SettingsOption_Supliers(){
                                 </div>
                                 
                                 <div className="btn-group" role="group">
-                                    <button type="button" className="btn btn-outline-danger" disabled={itemSelected===""?true:false} onClick={deleteUser}>Eliminar</button>     
-                                    <button type="button" className="btn btn-outline-info" disabled={itemSelected===""?true:false}>Editar</button>
-                                    {action===""?<button type="button" className="btn btn-outline-primary" onClick={insertUser}>Nuevo</button>:
+                                    <button type="button" className="btn btn-outline-danger" disabled={itemSelected==="" || action===`${process.env.REACT_APP_ACTIONS_MODIFY_SUPLIER}`?true:false} onClick={deleteSuplier}>Eliminar</button>     
+                                    <button type="button" className="btn btn-outline-info" disabled={itemSelected==="" || action===`${process.env.REACT_APP_ACTIONS_MODIFY_SUPLIER}`?true:false} onClick={editSuplier}>Editar</button>
+                                    {action===""?<button type="button" className="btn btn-outline-primary" onClick={insertSuplier}>Nuevo</button>:
                                                  <button type="button" className="btn btn-outline-primary" disabled={!validSuplier} onClick={submit}>Aceptar</button>}
                                 </div>
                             </div>
@@ -184,7 +205,7 @@ function SettingsOption_Supliers(){
                 </div>
             </div> 
 
-            <SubmitPassword action={action} onSuccess={submitSuccessful} onFailure={passwordFailure} onClose={paswordClose} parent="#SupliersModal" passId={(id)=>setPasswordId(id)}/>
+            <SubmitPassword action={action} onSuccess={submitSuccessful} onFailure={submitFailure} onClose={paswordClose} parent="#SupliersModal" passId={(id)=>setPasswordId(id)}/>
         </>
         
     );
